+++ tests/test_summary_generators.py
